+++ bosch-indego-controller-srvmock/src/main/java/de/zazaz/iot/bosch/indego/srvmock/IndegoServerMock.java
@@ -1,0 +1,13 @@
+package de.zazaz.iot.bosch.indego.srvmock;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IndegoServerMock {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(IndegoServerMock.class, args);
+    }
+
+}