+++ bosch-indego-controller-srvmock/src/main/java/de/zazaz/iot/bosch/indego/srvmock/service/AuthenticationService.java
@@ -1,0 +1,7 @@
+package de.zazaz.iot.bosch.indego.srvmock.service;
+
+public interface AuthenticationService {
+    
+    public String authenticate(String username, String password);
+
+}