+++ bosch-indego-controller-srvmock/src/main/java/de/zazaz/iot/bosch/indego/srvmock/service/ContextHandlerService.java
@@ -1,0 +1,9 @@
+package de.zazaz.iot.bosch.indego.srvmock.service;
+
+public interface ContextHandlerService {
+    
+    IndegoContext createContextForUser (String userId) throws IndegoServiceException;
+    
+    void deleteContext (String contextId) throws IndegoServiceException;
+
+}