+++ bosch-indego-controller-srvmock/src/main/java/de/zazaz/iot/bosch/indego/srvmock/service/IndegoContext.java
@@ -1,0 +1,11 @@
+package de.zazaz.iot.bosch.indego.srvmock.service;
+
+public interface IndegoContext {
+
+    String getUserId ();
+
+    String getDeviceSerial ();
+
+    String getContext ();
+
+}