+++ bosch-indego-controller-srvmock/src/main/java/de/zazaz/iot/bosch/indego/srvmock/service/IndegoService.java
@@ -1,0 +1,9 @@
+package de.zazaz.iot.bosch.indego.srvmock.service;
+
+import de.zazaz.iot.bosch.indego.DeviceStateInformation;
+
+public interface IndegoService {
+
+    DeviceStateInformation getState (String contextId, String serial) throws IndegoServiceException;
+
+}